[REFACTOR] 미들웨어 구성 변경
user 미들웨어를 의도에 맞게 사용하기 위해 미들웨어 구성 변경
