+++ server/models/query.js
@@ -5,6 +5,7 @@ module.exports = {
   DELETE_COMMENT: 'delete from comment where id = ?;',
   
   CREATE_USER: 'insert into user(is_auth, username, password) values(?, ?, ?);',
+  READ_ALL_USER: 'select id, is_auth, username, password from user;',
   READ_USER: 'select id, is_auth, username, password from user where username = ?;',
 
   CREATE_LABEL: 'insert into label(name, description, color) values(?, ?, ?);',
