+++ server/routes/user.js
@@ -1,4 +1,5 @@
 const express = require('express');
+const tokenMiddleWare = require('../middlewares/auth/token');
 const userMiddleWare = require('../middlewares/user');
 const passport = require('passport');
 
@@ -6,6 +7,8 @@ const router = express.Router();
 
 router.get('/auth/github/callback', passport.authenticate('github', {
   failureRedirect: '/',
-}), userMiddleWare.createToken);
+}), tokenMiddleWare.createToken);
+
+router.get('/all', userMiddleWare.readAllUser);
 
 module.exports = router;
